organizo los modelos en un modulo
